uint64_t -> uintptr_t

+++ Python/jit.c
@@ -116,7 +116,7 @@ mark_executable(unsigned char *memory, size_t size)
 
 // value[value_start : value_start + len]
 static uint32_t
-get_bits(uint64_t value, uint8_t value_start, uint8_t width)
+get_bits(uintptr_t value, uint8_t value_start, uint8_t width)
 {
     assert(width <= 32);
     return (value >> value_start) & ((1ULL << width) - 1);
@@ -124,7 +124,7 @@ get_bits(uint64_t value, uint8_t value_start, uint8_t width)
 
 // *loc[loc_start : loc_start + width] = value[value_start : value_start + width]
 static void
-set_bits(uint32_t *loc, uint8_t loc_start, uint64_t value, uint8_t value_start,
+set_bits(uint32_t *loc, uint8_t loc_start, uintptr_t value, uint8_t value_start,
          uint8_t width)
 {
     assert(loc_start + width <= 32);
@@ -295,7 +295,7 @@ patch_aarch64_21rx(unsigned char *location, uintptr_t value)
     //     // There should be only one register involved:
     //     assert(reg == get_bits(loc32[1], 0, 5));  // ldr's output register.
     //     assert(reg == get_bits(loc32[1], 5, 5));  // ldr's input register.
-    //     uint64_t relaxed = *(uint64_t *)value;
+    //     uintptr_t relaxed = *(uintptr_t *)value;
     //     if (relaxed < (1UL << 16)) {
     //         // adrp reg, AAA; ldr reg, [reg + BBB] -> movz reg, XXX; nop
     //         loc32[0] = 0xD2800000 | (get_bits(relaxed, 0, 16) << 5) | reg;
@@ -346,7 +346,7 @@ patch_x86_64_32rx(unsigned char *location, uintptr_t value)
 {
     uint8_t *loc8 = (uint8_t *)location;
     // Try to relax the GOT load into an immediate value:
-    uint64_t relaxed = *(uint64_t *)(value + 4) - 4;
+    uintptr_t relaxed = *(uintptr_t *)(value + 4) - 4;
     if ((int64_t)relaxed - (int64_t)location >= -(1LL << 31) &&
         (int64_t)relaxed - (int64_t)location + 1 < (1LL << 31))
     {
